# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Always_date_api documentation",
        "description": "Always date documentation for back-end",
        "contact": {
            "name": "FlokY#0101"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/life_school": {
            "get": {
                "tags": [
                    "Life school"
                ],
                "summary": "Iegūst visus life school, tikai pēc gender",
                "description": "Iegūst visas dzīves skolas rakstus, bet atgriežot dzīves skolas tiek filtrētas pēc lietotāja dzimuma",
                "operationId": "getLifeSchool",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Life school"
                ],
                "summary": "Izveido jaunu ierakstu par life school",
                "description": "Izveido jaunu ierakstu life school",
                "operationId": "postLifeSchool",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/LifeSchoolRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nevar noteikt dzimumu."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/life_school/{id}": {
            "get": {
                "tags": [
                    "Life school"
                ],
                "summary": "Atgriež dzīves skolas rakstu pēc id",
                "description": "Atgriež dzīves skolas rakstu pēc id",
                "operationId": "getLifeSchoolById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchool id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nedrīkst skatīt cita dzimuma dzīves skolu"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Life school"
                ],
                "summary": "Atjaunina datus dzīves skolai",
                "description": "Atjauno datus dzīves skolai pēc ievadītā id iekš uri",
                "operationId": "updateLifeSchool",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchool id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#components/schemas/LifeSchoolRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Life school"
                ],
                "summary": "Izdzēš ārā dzīves skolas rakstu",
                "description": "Dzēš ārā dzīves skolas rakstu pēc id",
                "operationId": "deleteLifeSchool",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "LifeSchool id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LifeSchoolResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LifeSchoolResource": {
                "title": "Life school resource",
                "description": "Life school resource",
                "xml": {
                    "name": "LifeSchoolResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}